Pull Request 1
S button, makes the player move faster. The space acts bar acts in 2 ways. In the upper part of the screen acts as  jump and in the second part of the screen acts as teleport and hide. It teleports the player after the border of the upper part, in this way hiding him from the bombs.